requirements.txt
typer
loguru
tqdm
streamlit
streamlit-folium
folium
shapely
pyproj
pandas
geopandas
plotly
prophet
python-dotenv
fsspec
webdav4[fsspec]  # filesystem
universal_pathlib  # filesystem
-e .